jobs/forms.py
from django import forms
from django.contrib.auth.models import User
from django.contrib.auth.forms import UserCreationForm
from ..users.models import Job
from django.db import transaction
import googlemaps
from crispy_forms.helper import FormHelper
from crispy_forms.layout import Layout, Submit, Row, Column

