requirements.txt
Django==2.1.7
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
Pillow==5.4.1
pytz==2018.9
